+++ thetatauCMT/scores/models.py
@@ -30,8 +30,9 @@ class ScoreType(models.Model):
         max_length=3,
         choices=[x.value for x in SECTION]
     )
-    points = models.PositiveIntegerField(default=0,
-                               help_text="Total number of points possible in year")
+    points = models.PositiveIntegerField(
+        default=0,
+        help_text="Total number of points possible in year")
     formula = models.CharField(max_length=200,
                                help_text="Formula for calculating score")
     name_short = models.CharField(max_length=20)
@@ -39,13 +40,13 @@ class ScoreType(models.Model):
         max_length=3,
         choices=[x.value for x in TYPES]
     )
-    base_points = models.PositiveIntegerField(default=0)
-    attendance_multiplier = models.PositiveIntegerField(default=0)
-    member_add = models.PositiveIntegerField(default=0)
+    base_points = models.FloatField(default=0)
+    attendance_multiplier = models.FloatField(default=0)
+    member_add = models.FloatField(default=0)
     special = models.CharField(max_length=200)
 
 
 class ScoreChapter(YearTermModel):
     chapter = models.ForeignKey(Chapter, related_name="scores")
     type = models.ForeignKey(ScoreType, related_name="chapters")
-    score = models.PositiveIntegerField(default=0)
+    score = models.FloatField(default=0)
